backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
fairlearn==0.10.0
shap==0.45.1
google-generativeai==0.7.2
reportlab==4.2.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0